% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stops.R
\name{find_stops}
\alias{find_stops}
\title{Find Stops}
\usage{
find_stops(latitude, longitude, datetime, stop_min_duration_s = 10,
  digits = 3, k = 3)
}
\arguments{
\item{latitude}{(numeric) GPS Latitude.}

\item{longitude}{(numeric) GPS Longitude.}

\item{datetime}{(datetime) GPS Timestamp.}

\item{stop_min_duration_s}{(integer) Minimum stop duration cutoff in seconds.}

\item{digits}{(integer) Decimal places used for rounding.}

\item{k}{(integer) Window length for rolling median. (Must be odd.)}
}
\description{
Find stops made in a route given coordinate and timestamps for a GPS track.
}
\examples{
find_stops(latitude, longitude, datetime)
}
\keyword{GPS,}
\keyword{stops}
