% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stops.R
\name{find_stops}
\alias{find_stops}
\title{Find Stops}
\usage{
find_stops(latitude, longitude, datetime, stop_min_duration_s = 10,
  digits = 3, k = 3)
}
\arguments{
\item{latitude}{(numeric) GPS Latitude.}

\item{longitude}{(numeric) GPS Longitude.}

\item{datetime}{(datetime) GPS Timestamp.}

\item{stop_min_duration_s}{(integer) Minimum stop duration cutoff in seconds.}

\item{digits}{(integer) Decimal places used for rounding.}

\item{k}{(integer) Window length for rolling median. (Must be odd.)}
}
\value{
(tibble) The coordinates and times of the identified stops, as a
    subset of the input data.
}
\description{
Find stops made in a route given coordinate and timestamps from a GPS track.
}
\section{Details}{

Stops are determined by finding sequential observations near the same location.
A rounded rolling median is used for latitude and longitude with \code{rle()}
to reduce noise. For this purpose, rounding is to \code{digits} decimal places.
The stop duration cutoff \code{stop_min_duration_s} is in seconds, assuming
one observation per second. Otherwise, consider this as the minimum number of
observations per stop.
}

\examples{
find_stops(latitude, longitude, datetime)
}
\keyword{GPS,}
\keyword{stops}
